requirements.txt
mlflow-skinny
seaborn
matplotlib
streamlit
pandas
joblib
scikit-learn
pytest